requirements.txt
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
cycler==0.10.0
Cython==0.28.1
decorator==4.2.1
dlib==19.13.1
enum34==1.1.6
h5py==2.7.1
imutils==0.4.6
ipython==5.6.0
ipython-genutils==0.2.0
kiwisolver==1.0.1
leveldb==0.194
matplotlib==2.2.2
networkx==2.1
nose==1.3.7
numpy==1.14.3
opencv-contrib-python==3.4.1.15
pandas==0.22.0
pathlib2==2.3.0
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.1.0
progressbar2==3.38.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pydot==1.2.4
Pygments==2.2.0
pyparsing==2.2.0
Pyste==0.9.10
python-dateutil==1.5
python-gflags==3.1.2
python-utils==2.3.0
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
scandir==1.7
scikit-image==0.13.1
scipy==0.17.0
simplegeneric==0.8.1
six==1.11.0
subprocess32==3.2.7
traitlets==4.3.2
wcwidth==0.1.7
